+++ vinisto/main.py
@@ -8,6 +8,7 @@ import inspect
 import logging
 import pkgutil
 import argparse
+from collections import defaultdict
 import vinisto.plugins
 import vinisto.tts
 import vinisto.stt
@@ -16,6 +17,14 @@ from importlib import import_module
 
 logging.basicConfig(level=logging.DEBUG)
 LOG = logging.getLogger(__name__)
+
+KEYWORDS = defaultdict(
+    lambda: ["vinisto", "Yes, master?"],
+    {
+        'es-ES': ['bautista', 'A sus ordenes'],
+        'en-US': ['Alfred', 'Yes, master?']
+    }
+)
 
 
 class Vinisto(object):
@@ -36,18 +45,24 @@ class Vinisto(object):
 
     def wait_for_keyword(self, keyword):
         while True:
-            LOG.info("Wating for keyword")
-            if next(self.stt.text) == keyword:
-                break
+            LOG.info("Wating for keyword {}".format(keyword))
+            text = next(self.stt.text)
 
-    def execute_callbacks(self):
+            if text == keyword:
+                yield False
+            elif keyword in text:
+                yield text.replace(keyword, '')
+
+    def execute_callbacks(self, text=False):
         """
             Launch callback function on all registered plugins
+            We can provide a text to be evaluated.
         """
-        try:
-            text = next(self.stt.text)
-        except LookupError:
-            pass
+        if not text:
+            try:
+                text = next(self.stt.text)
+            except LookupError:
+                pass
         for plugin in self.plugins:
             try:
                 LOG.info('Calling plugin {} callback'.format(plugin))
@@ -109,25 +124,49 @@ def main():
                         default='vinisto.tts.google_tts')
 
     parser.add_argument('--keyword',  type=str,
-                        help='Keyword to wait for',
-                        default='vinisto')
+                        help='Keyword to wait for')
+
+    parser.add_argument('--rate', type=str,
+                        help="Mic rate, defaults to raspberry pi USB (24000)",
+                        default=24000)
+
+    parser.add_argument('--language', type=str,
+                        help="Language to use in both TTS and STT",
+                        default="es-ES")
+
+    parser.add_argument('--key', type=str,
+                        help="Key to be passed to STT engines.",
+                        default="AIzaSyCuOvb2qd0mhQRkIbGAcgMUmFQaLIXtlmg")
+
+    parser.add_argument('--response_phrase', type=str,
+                        help="What to say when the keyword has been detected")
 
     args = parser.parse_args()
 
     TTS = next(extract_classes([args.tts]))
     STT = next(extract_classes([args.stt]))
 
-    vinisto_.stt = STT(language='es-ES', rate=24000,
-                       key="AIzaSyCuOvb2qd0mhQRkIbGAcgMUmFQaLIXtlmg")
-    vinisto_.tts = TTS(language='es-ES')
+    vinisto_.stt = STT(language=args.language, rate=args.rate,
+                       key=args.key)
+    vinisto_.tts = TTS(language=args.language)
 
     for class_ in extract_classes(plugins):
         vinisto_.register_plugin(class_)
 
-    while True:
-        vinisto_.wait_for_keyword(args.keyword)
-        LOG.info("Keyword recognized")
-        vinisto_.execute_callbacks()
+    if args.keyword:
+        keyword = args.keyword
+    else:
+        keyword = KEYWORDS[args.language][0]
+
+    if args.phrase:
+        phrase = args.phrase
+    else:
+        phrase = KEYWORDS[args.language][0])
+
+    for text in vinisto_.wait_for_keyword(keyword):
+        if not text:
+            vinisto_.tts.say(phrase)
+        vinisto_.execute_callbacks(text)
 
 
 if __name__ == "__main__":